Write types in CommonJS, so they can be exported from ESM files
